% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\docType{data}
\name{animal_sleep}
\alias{animal_sleep}
\title{This dataset contains the details 62 animals and their various attributes}
\format{
An object of class \code{data.frame} with 62 rows and 11 columns.
}
\usage{
animal_sleep
}
\description{
This dataset contains the details 62 animals and their various attributes
}
\keyword{datasets}
